Bug 795720 - Change MD5 sums to SHA512 ones.
Keep the MD5 sums in the GIMP_VERSIONS database, and add the sha256 and
sha512 sums when available (i.e. only for the 2.10.0 release). The
website code will now prefer sha256 when available, then sha512 and
finally fallback to md5 as last resort.
The reason why I don't set sha512 as first choice yet is because we'd
want to have some CSS for the output to not be very ugly (the sums are
also available in the download server anyway). Once we have some
pretty-printing CSS for these long hash sums, let's switch to SHA512.