Bug 795709 - Flatpack as only option is not enough.
Reorder a bit the text:
(1) Emphasize a bit the fact that flatpak is *NOT* the recommended way
of installation. Distribution packages still are prefered when possible,
if not for anything else, because flatpak still has a few issues and
even some feature loss relative to the sandbox model. Flatpak is an
additional installation medium, with its own advantages, but is still
not about to replace totally distribution packing systems.
(2) Add a title to the additional flatpak instructions to make them
stand out a bit more, in the case when the "easy way" does not work
(i.e. when flatpak is not well integrated in the distribution).
(3) Add a title for the other OSes, not supported by Flatpak, to also
make them stand out a bit more, and not having them too much lost at the
end of a wall of text.