Sync OLDSTABLE stuff in pelican.testing to pelican.local
schumaml started adding this to support an OLDSTABLE download page
for people looking for 2.8 stuff.  This is just syncing his work
to the pelican.local.py file.

I am not sure we need an entire download page just for the old
2.8 stuff - I feel like a simple set of links to the mirror to
get the file should be enough.

I've changed downloads.html template to add these links
automatically.  If we don't like it, we can still work on the
OLDSTABLE page for 2.8.